requirements.txt
pandas==1.5.3
openpyxl==3.1.2
streamlit==1.25.0
langchain
langchain-google-genai
google-auth==2.23.4
pyyaml==6.0.1
pydantic==2.1.1  
